Typo fixes; fixed added check for successful build before compressing

+++ source_files/obsidian_main/g_doom.cc
@@ -938,32 +938,34 @@ bool Doom::game_interface_c::Finish(bool build_ok) {
         Recent_AddFile(RECG_Output, filename);
     }
 
-    if (zip_output > 0) {
-        std::filesystem::path zip_filename = filename;
-        zip_filename.replace_extension(zip_output == 1 ? "zip" : "pk3");
-        if (std::filesystem::exists(zip_filename)) {
-            if (create_backups) {
-                Main::BackupFile(zip_filename, ".old");
+    if (build_ok) {
+        if (zip_output > 0) {
+            std::filesystem::path zip_filename = filename;
+            zip_filename.replace_extension(zip_output == 1 ? "zip" : "pk3");
+            if (std::filesystem::exists(zip_filename)) {
+                if (create_backups) {
+                    Main::BackupFile(zip_filename, ".old");
+                }
+                std::filesystem::remove(zip_filename);
             }
-            std::filesystem::remove(zip_filename);
-        }
-        FILE *zip_file = fopen(filename.string().c_str(), "rb");
-        int zip_length = std::filesystem::file_size(filename);
-        byte *zip_buf = new byte[zip_length];
-        if (zip_buf && zip_file){
-            memset(zip_buf, 0, zip_length);
-            fread(zip_buf, 1, zip_length, zip_file);
-        }
-        if (zip_file) fclose(zip_file);
-        if (zip_buf) {
-            if (mz_zip_add_mem_to_archive_file_in_place(zip_filename.string().c_str(), filename.filename().string().c_str(), zip_buf, zip_length, NULL, 0, MZ_DEFAULT_COMPRESSION)) {
-                std::filesystem::remove(filename);
-                delete []zip_buf;
+            FILE *zip_file = fopen(filename.string().c_str(), "rb");
+            int zip_length = std::filesystem::file_size(filename);
+            byte *zip_buf = new byte[zip_length];
+            if (zip_buf && zip_file){
+                memset(zip_buf, 0, zip_length);
+                fread(zip_buf, 1, zip_length, zip_file);
+            }
+            if (zip_file) fclose(zip_file);
+            if (zip_buf) {
+                if (mz_zip_add_mem_to_archive_file_in_place(zip_filename.string().c_str(), filename.filename().string().c_str(), zip_buf, zip_length, NULL, 0, MZ_DEFAULT_COMPRESSION)) {
+                    std::filesystem::remove(filename);
+                    delete []zip_buf;
+                } else {
+                    LogPrintf("Zipping output WAD to {} failed! Retaining original WAD.\n", zip_filename.generic_string());
+                }
             } else {
                 LogPrintf("Zipping output WAD to {} failed! Retaining original WAD.\n", zip_filename.generic_string());
             }
-        } else {
-            LogPrintf("Zipping output WAD to {} failed! Retaining original WAD.\n", zip_filename.generic_string());
         }
     }
